% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{camelCaseToSnakeCase}
\alias{camelCaseToSnakeCase}
\title{Convert a camel case string to snake case}
\usage{
camelCaseToSnakeCase(string)
}
\arguments{
\item{string}{The string to be converted}
}
\value{
A string
}
\description{
Convert a camel case string to snake case
}
\examples{
camelCaseToSnakeCase("exposureConceptId1")

}
