% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{camelCaseToSnakeCaseNames}
\alias{camelCaseToSnakeCaseNames}
\title{Convert the names of an object from camel case to snake case}
\usage{
camelCaseToSnakeCaseNames(object)
}
\arguments{
\item{object}{The object of which the names should be converted}
}
\value{
The same object, but with converted names.
}
\description{
Convert the names of an object from camel case to snake case
}
\examples{
x <- data.frame(conceptId = 1, conceptName = "b")
camelCaseToSnakeCaseNames(x)

}
