% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{camelCaseToTitleCase}
\alias{camelCaseToTitleCase}
\title{Convert a camel case string to title case}
\usage{
camelCaseToTitleCase(string)
}
\arguments{
\item{string}{The string to be converted}
}
\value{
A string
}
\description{
Convert a camel case string to title case
}
\examples{
camelCaseToTitleCase("exposureConceptId1")

}
