% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderSql.R
\name{getTempTablePrefix}
\alias{getTempTablePrefix}
\title{Get the prefix used for emulated temp tables for DBMSs that do not support temp tables (e.g. Oracle,
BigQuery).}
\usage{
getTempTablePrefix()
}
\value{
The prefix string.
}
\description{
Get the prefix used for emulated temp tables for DBMSs that do not support temp tables (e.g. Oracle,
BigQuery).
}
\examples{
getTempTablePrefix()
}
