% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{launchSqlDeveloper}
\alias{launchSqlDeveloper}
\title{Launch the SQL Developer Shiny app}
\usage{
launchSqlDeveloper()
}
\description{
Launch the SQL Developer Shiny app
}
\details{
Launches a Shiny app that allows the user to develop SQL and see how it translates to the supported dialects.
}
