% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{launchSqlRenderDeveloper}
\alias{launchSqlRenderDeveloper}
\title{Launch the SqlRender Developer Shiny app}
\usage{
launchSqlRenderDeveloper(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the SqlRender Developer Shiny app
}
\details{
Launches a Shiny app that allows the user to develop SQL and see how it translates to the supported dialects.
}
