% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{readSql}
\alias{readSql}
\title{Reads a SQL file}
\usage{
readSql(sourceFile)
}
\arguments{
\item{sourceFile}{The source SQL file}
}
\value{
Returns a string containing the SQL.
}
\description{
\code{readSql} loads SQL from a file
}
\details{
\code{readSql} loads SQL from a file
}
\examples{
\dontrun{
readSql("myParamStatement.sql")
}
}
