% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderSql.R
\name{renderSql}
\alias{renderSql}
\title{Deprecated: Render SQL code based on parameterized SQL and parameter values}
\usage{
renderSql(sql = "", warnOnMissingParameters = TRUE, ...)
}
\arguments{
\item{sql}{The parameterized SQL}

\item{warnOnMissingParameters}{Should a warning be raised when parameters provided to this function 
do not appear in the parameterized SQL that is being rendered? By default, this is TRUE.}

\item{...}{Parameter values}
}
\value{
A list containing the following elements: \describe{ \item{parameterizedSql}{The original
parameterized SQL code} \item{sql}{The rendered sql} }
}
\description{
This function has been deprecated. Use \code{\link{render}} instead. This new 
function returns a character vector instead of a list.
}
