% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{snakeCaseToCamelCase}
\alias{snakeCaseToCamelCase}
\title{Convert a snake case string to camel case}
\usage{
snakeCaseToCamelCase(string)
}
\arguments{
\item{string}{The string to be converted}
}
\value{
A string
}
\description{
Convert a snake case string to camel case
}
\examples{
snakeCaseToCamelCase("cdm_database_schema")
# > 'cdmDatabaseSchema'

}
