% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{snakeCaseToCamelCaseNames}
\alias{snakeCaseToCamelCaseNames}
\title{Convert the names of an object from snake case to camel case}
\usage{
snakeCaseToCamelCaseNames(object)
}
\arguments{
\item{object}{The object of which the names should be converted}
}
\value{
The same object, but with converted names.
}
\description{
Convert the names of an object from snake case to camel case
}
\examples{
x <- data.frame(concept_id = 1, concept_name = "b")
snakeCaseToCamelCaseNames(x)

}
