% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{supportsJava8}
\alias{supportsJava8}
\title{Determine if Java virtual machine supports Java}
\usage{
supportsJava8()
}
\value{
Returns TRUE if JVM supports Java >= 8.
}
\description{
Tests Java virtual machine (JVM) java.version system property to check if version >= 8.
}
\examples{
supportsJava8()
}
