% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{writeSql}
\alias{writeSql}
\title{Write SQL to a SQL (text) file}
\usage{
writeSql(sql, targetFile)
}
\arguments{
\item{sql}{A string containing the sql}

\item{targetFile}{The target SQL file}
}
\description{
\code{writeSql} writes SQL to a file
}
\details{
\code{writeSql} writes SQL to a file
}
\examples{
\dontrun{
sql <- "SELECT * FROM @table_name"
writeSql(sql, "myParamStatement.sql")
}
}
