% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderSql.R
\name{sparkHandleInsert}
\alias{sparkHandleInsert}
\title{Handle Spark insert statements}
\usage{
sparkHandleInsert(connection, sql)
}
\arguments{
\item{connection}{The connection to the database server.}

\item{sql}{The SQL to be executed}
}
\description{
Handle Spark insert statements
}
