% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlRender.R
\docType{package}
\name{SqlRender-package}
\alias{SqlRender}
\alias{SqlRender-package}
\title{SqlRender: Rendering Parameterized SQL and Translation to Dialects}
\description{
A rendering tool for parameterized SQL that also translates into different SQL dialects. These dialects include 'Microsoft SQL Server', 'Oracle', 'PostgreSql', 'Amazon RedShift', 'Apache Impala', 'IBM Netezza', 'Google BigQuery', 'Microsoft PDW', 'Snowflake', 'Azure Synapse Analytics Dedicated', 'Apache Spark', 'SQLite', and 'InterSystems IRIS'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/SqlRender/}
  \item \url{https://github.com/OHDSI/SqlRender}
  \item Report bugs at \url{https://github.com/OHDSI/SqlRender/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Marc Suchard
}

}
\keyword{internal}
